remove tests with deprecated code
